A ball, and a floor! Yeay!

+++ core/src/com/teamkarbon/android/test_gdx/Ball.java
@@ -1,15 +1,15 @@
 package com.teamkarbon.android.test_gdx;
 
 public class Ball {
-    int x;
-    int y;
+    float x;
+    float y;
 
-    public Ball(int _x, int _y){
+    public Ball(float _x, float _y){
         x = _x;
         y = _y;
     }
 
-    public void setPos(int xpos, int ypos)
+    public void setPos(float xpos, float ypos)
     {
         x = xpos;
         y = ypos;
